remove withStickyPlayer from App.jsx
